+++ src/main/java/wooteco/subway/station/controller/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station.dto;
+package wooteco.subway.station.controller.dto;
 
 public class StationRequest {
     private String name;
